Update patch set 2
Patch Set 2:

(1 comment)

This is much better than what we have now, although I still wonder if we can get a little less granular here. I think the concept of "transitioning" between different states might be confusing to users. It seems like we might want to be a little more opaque about the states of the state machine here, so we indicate that we can't read a document while positioned on a Null or that we can't read the next element while positioned on an element.

Patch-set: 2
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>