Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Looking at https://jira.mongodb.org/browse/DRIVERS-290, it implies that write concern is only supported on server >= 3.4.
> 
> Also, https://github.com/mongodb/specifications/blob/master/source/read-write-concern/read-write-concern.rst#servers-default-writeconcern says that the default write concern should not be serialized.  I'm not sure if the default write concern is represented as nil in the Go driver (Java and .NET it is not, for example).  Can you confirm?  If not, then we need to open an issue about it.

Fixed the wire version issue in the new patch.
We do incorrectly encode the default write concern in the command. If a collection is created with the following opts:
options.Collection().SetWriteConcern(writeconcern.New()),
a write concern with a value of empty doc will be sent to the server for write commands.

Patch-set: 2
Reviewer: Divjot Arora <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>