Update patch set 11
Patch Set 11:

(24 comments)

> Patch Set 11:
> 
> (25 comments)
> 
> Looks good. Couple of comments/questions apart from those inline:
> 
> 1. In the decoders, it seems like we need to be checking for bson null everywhere... not just in a few places. Am I missing something in, for instance, IntDecodeValue method that will handle this?
> 
> 2. I don't see a lot of new tests related to the bugs that are getting fixed. Can you point those out?

There are only a few places where BSON Null is valid, mainly types that are nialble in Go. For example, we cannot set an int64 to nil, so attempting to decode BSON Null into it is an error.

I've added a number of new tests, and there were already many success path tests. Because of how I wrote the test case runners, the runners for the error cases didn't work well for the plethora of success cases, so those are in their own subtest.

For the things that are fixed, I haven't pushed the new patch yet but will by end of day.

Patch-set: 11
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>