Update patch set 1
Patch Set 1:

(1 comment)

This might be difficult in Evergreen, but would it be possible to have a test to show that this fixes the issue described in https://jira.mongodb.org/browse/GODRIVER-641?

If this isn't possible, maybe we could at create a client, encode a command, and ensure that slaveOK bit is set (or $readPreference is encoded for OP_MSG)?

Patch-set: 1
CC: Divjot Arora <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>