Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> This might be difficult in Evergreen, but would it be possible to have a test to show that this fixes the issue described in https://jira.mongodb.org/browse/GODRIVER-641?
> 
> If this isn't possible, maybe we could at create a client, encode a command, and ensure that slaveOK bit is set (or $readPreference is encoded for OP_MSG)?

A test like that seems like it'll be either rely on too many mocks (basically make a fake topology and try to use it to process an actual command) or require a lot of orchestration (connect to a replicaset, determine which members are secondaries, connect to them in single mode, and verify that the commands we send are correct). Both those options seem brittle.

Patch-set: 1
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>