Update patch set 1
Patch Set 1:

Because of gerrit issues, I had to start a new CR rather than push a new patch, but this is mainly based on the old patch. This patch is built on top of the patch for closing implicit sessions if errors occur when running cursor commands. The new changes:

1. x/mongo/driver/topology/cursor.go changed to remove the legacy flag from the cursor struct. The cursor chooses between killCurors/OP_KILL_CURSORS and getMore/OP_GET_MORE based on the server description.

2. New tests added to test errors for OP_QUERY and find in mongo/collection_internal_test.go

3. x/mongo/driver/find.go and x/network/wiremessage/query.go changed to pass negative limits directly to the OP_QUERY struct instead of making them positive so the command monitoring upconversion in query.go can correctly add the singleBatch field to the command document.

Patch-set: 1
Reviewer: Divjot Arora <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>