Update patch set 4
Patch Set 4:

(7 comments)

Thanks for the feedback Divjot and Kris! In addition to the control flow changes, I switched to table-driven tests and made the test names shorter as suggested. I decided to put the OP_MSG and OP_QUERY tests in separate test tables because the way I was testing them was different enough that I had to include a bunch of if statements that controlled what we tested. That said, let me know if you disagree with this decision or have a better approach.

Patch-set: 4
Reviewer: Devin Hilly <1015552@d5d70762-12d0-45a1-890d-524b12d3f735>