Update patch set 2
Patch Set 2:

Talked to Kris offline and we decided that changing SSLOpt to a TLSConfig requires too many changes to the connstring parsing code in topology_options.go. For now, we've decided to leave this as a bug fix and I've created GODRIVER-747 as an improvement.

Patch-set: 2
Reviewer: Gerrit User 1014410 <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>