Update patch set 3
Patch Set 3:

(2 comments)

I'm a little worried about how much work we'll have to do for the changes required by SPEC-1122, but I don't think this patch is the place to do it.

For this patch specifically, it looks like we'll attempt to resume forever when we get a resumable error instead of just once as the specification indicates.

Patch-set: 3
Reviewer: Gerrit User 1001503 <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>