Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> I was looking for a test that could be modified to test the new field, and found  mongo/crud_util_test.go.  Is it necessary or useful to add Normalization to collationFromMap and/or newCollationFromMap (not sure why there are two functions that look identical)?  
> 
> Either way, seems like `func (co *Collation) ToDocument()` should have a unit test.

Not sure why those functions were identical either, but the comment on newCollationFromMap mentioned mongoopt, which is from the old options API and mentioned that the function could be deleted with the new options work, so I removed it and changed any references in the latest patch. Also added a unit test to ensure that the collation is serialized to a BSON document correctly.

Patch-set: 2
Reviewer: Gerrit User 1014410 <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>