Update patch set 1
Patch Set 1:

The point of contention on this ticket is whether or not we actually need an internal TLS config wrapper. If we remove the internal wrapper, we're removing the support to decrypt client certificates for users and requiring them to directly give us a tls.Config instance.

TODO: talk to Craig and David and reach a consensus on whether or not we want to keep a tls.Config wrapper (and where it belongs if we do) and what we want to do about decrypting client certs (current options are to not do it or to release an open source lib which we vendor as a dependency).

Patch-set: 1
Reviewer: Gerrit User 1014410 <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>