Update patch set 1
Patch Set 1:

(1 comment)

Based on my comment in change_stream.go, the current patch will likely cause tests to fail. This can be fixed by reverting the change_stream.go code to directly kill the cursor. This will work, but it will mean that we can't mock batch cursors for change streams (it will assume that the underlying batch cursor is always of type *driver.BatchCursor). If we think mocking batch cursors for change streams will be helpful in the future, we'd need to look into a different way of tracking when implicit sessions should be closed in cursors. Based on an offline convo with Jeff, the Java driver uses a reference counting mechanism to do this.

Patch-set: 1
Reviewer: Gerrit User 1014410 <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>