Improve documentation and make commands for tests
Previously the readme said that tests could be run with `make`.
This was not the case as a TOPOLOGY variable had to be set. This
patch adds a check to ensure that variables used in testing are
set to valid values. It also outlines usage of testing variables
in the readme.

Change-Id: I7f60fb78282f08ca528f64e40b87fed67b5780f7