Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Thanks for the PR.
> 
> Let's make sure to also update .evergreen/config.yml so this configuration is regression-tested.

Absolutely. However it looks like the driver tools don't have orchestration configs for zlib compression. I suggest that we update the driver tools and then update .evergreen/config.yml. 
I'm happy to take that on. 

Do we want to make sure that the updated evergreen config is part of this patch, or are we okay with that landing a little later?

Patch-set: 2
Reviewer: Gerrit User 1015492 <1015492@d5d70762-12d0-45a1-890d-524b12d3f735>