Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > Thanks for the PR.
> > 
> > Let's make sure to also update .evergreen/config.yml so this configuration is regression-tested.
> 
> Absolutely. However it looks like the driver tools don't have orchestration configs for zlib compression. I suggest that we update the driver tools and then update .evergreen/config.yml. 
> I'm happy to take that on. 
> 
> Do we want to make sure that the updated evergreen config is part of this patch, or are we okay with that landing a little later?

Can you clarify what would need to be done to update the tools? Also @jeff, do you know if any other drivers test with zlib on Evergreen? I'd like the Evergreen config update to be part of this patch if possible, but if it's a nontrivial amount of work, I'm okay with a separate ticket for it as well.

Patch-set: 2