Update patch set 2
Patch Set 2:

Thanks for the explanation.  Were you able to test this manually, while we wait for the SPEC ticket to test this to be resolved?  It doesn't seem like this could be worse that what's currently in master, but would be nice to know that it works!

Patch-set: 2
Reviewer: Gerrit User 1013848 <1013848@d5d70762-12d0-45a1-890d-524b12d3f735>