Update patch set 2
Patch Set 2:

Patchset 2 incorporates a large part of the driverx package from the prototype. This is a placeholder for the actual driver package's OperationContext and CommandOperation types. We need these now because we can't migrate to the new topology.connection type until we move the handshakers over, so we'll use CommandOperation for now.

Patch-set: 2
Reviewer: Gerrit User 1001503 <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>