Update patch set 2
Patch Set 2:

(3 comments)

> Patch Set 2:
> 
> Operation functions were changed to have a pointer receiver because functions like createWireMessage need to set fields on the operation to keep state for command monitoring.

I don't mind that we switched things to pointers, but as I said in my comment, I think that we are going about this wrong. I think we should either return up the chain an event.CommandStartedEvent or create our own type that we can convert to an event.CommandStartedEvent. I think we can handle the copying behavior if we are actually going to send the CommandStartedEvent to the user.

It might make more sense to create our own type that's returned up the stack, so that it can hold the batches and we have to do minimal copying in the case of command monitoring being off.

Patch-set: 2
Reviewer: Gerrit User 1001503 <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>