Update patch set 5
Patch Set 5:

(1 comment)

I'm not sure if this is something we already have, but it's something to look into. Because we explicitly ignore the error from endSessions, we have no idea if it's even sent to the server. We should have an integration test using command monitoring that opens/connects a client, does a bunch of operations using implicit sessions (this should be handled automatically if you just run an operation without a session), calls Disconnect, and verifies that an endSessions command is sent to the server.

Patch-set: 5
Reviewer: Gerrit User 1014410 <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>