Update patch set 2
Patch Set 2:

(19 comments)

Good work implementing this but I think there are some things we need to resolve between the way the spec is written and our design.

We shouldn't need to implement our own wait queue, the semaphore package we are using in the topology.Server type should be usable in the topology.pool type. I think we need to redesign some of the code paths and maybe drop the use of the resource pool because it's over complicating the code and I don't think it's giving us that much of a benefit. We also are using context.Background in places where the user should be able to pass down a context.Context and it doesn't look like we're handling errors everywhere we should be.

I think a useful next step would be for us to sit down offline and go over the spec and see if we can make an implementation with a less verbose design.

Patch-set: 2
Reviewer: Gerrit User 1001503 <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>