Update patch set 1
Patch Set 1:

Btw, I wasn't sure what the process is for fixing a bug in code in a recently-merged ticket for an ongoing project. For the core API, we sometimes made tickets and sometimes added fixes without ticket numbers. For this, I left the ticket num as 1250. That way, it'll show up as a comment on Jira and both patches will be tracked.

Patch-set: 1
Reviewer: Gerrit User 1014410 <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>