Update patch set 2
Patch Set 2:

PS 2 changes the test setup so that client options are only specified if the test is not the one about maxWireVersion. The issue is that the maxWireVersion test fails because mongocryptd cannot be spawned. This should be resolved by SPEC-1403, but until then, this hack is necessary.

Patch-set: 2
Reviewer: Gerrit User 1014410 <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>