Improve clarity of incomplete read error message
Previously, the error message for being unable to read a document was
confusing because it mentioned "decoding" when no such decoding was
occuring.  E.g.:

> unable to decode message length: read tcp 10.148.11.183:42418->63.35.244.142:27017: i/o timeout

This commit changes this and a related message to clarify that the
problem is an incomplete read, either of the message header or the full
message.  (These terms are also not quite accurate, but are likely to be
better-understood by end-users reading the message.)  After the change,
the message above would appear as follows:

> incomplete read of message header: read tcp 10.148.11.183:42418->63.35.244.142:27017: i/o timeout