GODRIVER-1615 Track wire version on ChangeStream
This was needed to fix test errors. Because the deployment was changed
and the new deployment closes the connection right after the aggregate,
any code that access the wire version after the aggregate finishes would
panic. To fix that, we store the wire version of the last used connection
on ChangeStream.