Refactor to (almost) not use bsonx package
Removed the cases where it checks/creates a bsonx.Doc, and replaced
them with bson.D type if necessary. Removed functions that seemed only
to transform bsonx to another type. Removed a test for an specific
function for the same reason.

The bsonx package is largely replaced by bson and bsoncore. bsonx
was a proof of concept, and it should be completely replaced.

There is still various places that use bsonx:

	- benchmark directory
	- mongo/gridfs/bucket.go
	- mongo/gridfs/upload_stream.go
	- mongo/integration/collection_test.go
	- mongo/integration/gridfs_spec_test.go

And a file inside the same bsonx package uses bsonx package:
	- x/bsonx/array.go