bsoncore.readLengthBytes handle invalid lengths
The length by definition needs to include the 4 bytes for the size.
If a size of zero is provided this can result in an infinite loop as providing the zero to `src[l:]` will result in the slice not being consumed.  Allowing the zero length to be read for infinity.