GODRIVER-3599: Add task script to generate CycloneDX SBOM
Added a `task` and `etc` bash script to generate a build lifecycle
CycloneDX SBOM using a pinned version of the `cyclonedx-gomod` tool. The
SBOM includes the aggregate of modules required by packages in the
mongo-go-driver library, excluding examples, tests and test packages.

The task will run only when go.mod is newer than sbom.cdx.json.

The file is saved as sbom.cdx.json (as opposed to the current sbom.json)
which is the preferred file extension for CycloneDX files. There is not
yet any code to commit the new SBOM to the repo. This is to allow for
evaluation of the new SBOM first without interfering with the current
workflow to upload the static sbom.json file to Kondukto.

TODOs:
- Once the SBOM generation process has been approved, the GitHub Action
in `mongodb-labs/drivers-github-tools` can be called with the updated
SBOM file name.
- Add libmongocrypt as an optional component via a merge once the
libmongocrypt SBOM is updated with newer automation

`task generate-sbom`